roundtable_dockerize/summarize_api/summarization/requirements.txt
click
h11
pydantic
starlette
fastapi
PyJWT
passlib
uvicorn
python-dotenv
transformers
torch
nltk